Refine the condition for reading AdditionalPlayerInfo.
Apparently, if a player only has placeholder heroes, the logic for reading/writing .h3m is
slightly different:
* Name and portrait of the starting hero should not be read/written.
* additional_info should be read/written even though the type of the starting hero is 0xFF.

I'm not sure if this is completely accurate now, but at least it fixes the behavior on the map
where I noticed the issue ("Oblivion's Edge" scenario from the "Armageddon's Blade" campaign after
converting it to SoD).

Keeping this in a separate branch for now.