Reinterpret PlayerSpecs.
OK, so in other H3M parsers PlayerSpecs always stores the name and the portrait
of the starting hero, but only stores additional_info if certain conditions are
met. This doesn't seem to be the case - if the type of the starting hero is
0xFF, then instead of storing the portrait and the name, the map stores
additional_info, but this is ambiguous because:
* num_placeholder_heroes has the same size as portrait (1 byte)
* heroes is a vector whose size is stored in uint32, which is the same as the
  data type used to store the length of the string.
* usually, if the type of the starting hero is 0xFF, then the name is an empty
  string, so the above-mentioned uint32 will store 0, and there will be no
bytes after it.

However, if the player only has placeholder heroes in the beginning of the
game, then the type of the starting hero *might* be 0xFF, yet specific
placeholder heroes (i.e. configured as "Specific Hero" rather than "Power
Rating") will still be written *instead of* the portrait and the name of the
starting hero. This interpretation is more clear than the hacks that other
parsers do, so I'm switching to it.