Add a second parameter (object subtype) to getMetaObjectType().
There are 2 objec types for which MetaObjectType cannot be deduced from ObjectClass alone:
ObjectClass::MINE and ObjectClass::ABANDONED_MINE. When subtype is within [0; 6] they
represent a regular mine, with the owner (PlayerColor) being the only property. However,
if subtype == 7, then they represent an Abandoned Mine instead, and the only property is
a ResourcesBitmask (which, by a lucky coincidence, has the same size as PlayerColor: 1 byte).

The Map Editor uses both MINE and ABANDONED_MINE for standard object templates representing an
Abandoned Mine. It never uses ABANDONED_MINE for a regular mine, but manually setting the subtype
to a value from [0; 6] works fine both in the Map Editor and in the game.